rename plugin from openid to oidc
