+++ gradle/artifacts/publishJvmAndCommonPublications.txt
@@ -384,13 +384,13 @@ io.ktor:ktor-server-auth-ldap/.jar
 io.ktor:ktor-server-auth-ldap/javadoc.jar
 io.ktor:ktor-server-auth-ldap/kotlin-tooling-metadata.json
 io.ktor:ktor-server-auth-ldap/sources.jar
-io.ktor:ktor-server-auth-openid-jvm/.jar
-io.ktor:ktor-server-auth-openid-jvm/javadoc.jar
-io.ktor:ktor-server-auth-openid-jvm/sources.jar
-io.ktor:ktor-server-auth-openid/.jar
-io.ktor:ktor-server-auth-openid/javadoc.jar
-io.ktor:ktor-server-auth-openid/kotlin-tooling-metadata.json
-io.ktor:ktor-server-auth-openid/sources.jar
+io.ktor:ktor-server-auth-oidc-jvm/.jar
+io.ktor:ktor-server-auth-oidc-jvm/javadoc.jar
+io.ktor:ktor-server-auth-oidc-jvm/sources.jar
+io.ktor:ktor-server-auth-oidc/.jar
+io.ktor:ktor-server-auth-oidc/javadoc.jar
+io.ktor:ktor-server-auth-oidc/kotlin-tooling-metadata.json
+io.ktor:ktor-server-auth-oidc/sources.jar
 io.ktor:ktor-server-auth/.jar
 io.ktor:ktor-server-auth/javadoc.jar
 io.ktor:ktor-server-auth/kotlin-tooling-metadata.json
