+++ settings.gradle.kts
@@ -63,7 +63,7 @@ projects {
             +"ktor-server-auth-api-key"
             +"ktor-server-auth-jwt"
             +"ktor-server-auth-ldap"
-            +"ktor-server-auth-openid"
+            +"ktor-server-auth-oidc"
             +"ktor-server-auto-head-response"
             +"ktor-server-body-limit"
             +"ktor-server-caching-headers"
